// src/types.ts
// src/types.ts
export interface Veteran {
    Id: number;
    Name: string;
    GivenNames: string;
    Surname: string;
    BirthDate: string;
    BirthLocation: string;
    Father: string;
    Mother: string;
    Siblings: string;
    Spouse: string;
    Children: string;
    OtherFamily: string;
    HighSchool: string;
    HighSchoolLocation: string;
    HighSchoolActivities: string;
    HomeTown: string;
    HomeState: string;
    EnteredKnox: string;
    ClassYear: string;
    Major: string;
    Sports: string;
    Honored: string;
    Activities: string;
    GreekOrganization: string;
    EnlistmentDate: string;
    EnlistmentLocation: string;
    EnlistmentNumber: string;
    War: string;
    Service: string;
    SerialNumber: string;
    Rank: string;
    Unit: string;
    ServiceHistory: string;
    Incident: string;
    IncidentDate: string;
    IncidentLocation: string;
    CasualtyData: string;
    YearOfDeath: string;
    DateOfDeath: string;
    AgeAtDeath: string;
    LocationOfDeath: string;
    OriginallyBuried: string;
    Burial: string;
    PlotLocation: string;
    CivilianCenotaph: string;
    MilitaryCenotaph: string;
    Links: string;
    FindAGrave: string;
    Awards: string;
    Citation: string;
    Remarks: string;
    KnoxID: string;
    LastUpdated: string;
    Photos: string;
}